+++ zmu_cortex_m/src/decoder/lsr.rs
@@ -39,7 +39,7 @@ pub fn decode_LSR_reg_t1(opcode: u16) -> Instruction {
 pub fn decode_LSR_imm_t2(opcode: u32) -> Instruction {
     let rm: u8 = opcode.get_bits(0..4) as u8;
     let rd: u8 = opcode.get_bits(8..12) as u8;
-    let s: u8 = opcode.get_bit(20) as u8;
+    let s: u8 = u8::from(opcode.get_bit(20));
 
     let imm3: u8 = opcode.get_bits(12..15) as u8;
     let imm2: u8 = opcode.get_bits(6..8) as u8;
@@ -66,13 +66,13 @@ pub fn decode_LSR_reg_t2(opcode: u32) -> Instruction {
     let rn: u8 = opcode.get_bits(16..20) as u8;
     let rm: u8 = opcode.get_bits(0..4) as u8;
     let rd: u8 = opcode.get_bits(8..12) as u8;
-    let s: u8 = opcode.get_bit(20) as u8;
+    let s: u8 = u8::from(opcode.get_bit(20));
 
     Instruction::LSR_reg {
         params: Reg3Params {
-            rd: Reg::from(rd as u8),
-            rn: Reg::from(rn as u8),
-            rm: Reg::from(rm as u8),
+            rd: Reg::from(rd),
+            rn: Reg::from(rn),
+            rm: Reg::from(rm),
             setflags: if s == 1 {
                 SetFlags::True
             } else {
